Append should keyword to test description #2
Description of the Change
Adding the should keyword to an example unit test.

Why should this be in core?
Docs are part of the core package.

Benefits
Maintain consistency of documentation.

Possible Drawbacks
No drawbacks.

Applicable issues
Is this a breaking change (major release)? No
Is it an enhancement (minor release)? No
Is it a bug fix, or does it not impact production code (patch release)? Does not impact production code.