Append should keyword to test description #3
Description of the Change
Adding the should keyword to an example unit test.

Why should this be in core?
Docs are part of the core package.

Benefits
Maintain consistency of documentation.

Possible Drawbacks
No drawbacks.

Applicable issues
Is this a breaking change (major release)? No
Is it an enhancement (minor release)? No
Is it a bug fix, or does it not impact production code (patch release)? Does not impact production code.

Requirements
Filling out the template is required. Any pull request that does not include enough information to be reviewed in a timely manner may be closed at the maintainers' discretion.
All new code requires tests to ensure against regressions.
Description of the Change
Alternate Designs
Why should this be in core?
Benefits
Possible Drawbacks
Applicable issues